\name{MBE}
\alias{MBE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Mean bias error
}
\description{
Calculates the normalized (w.r.t. mean) mean bias error
}
\usage{
MBE(meas, pred)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{meas}{Measured values}
  \item{pred}{Predicted values}
}

