\name{RMSE}
\alias{RMSE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Root mean squares error
}
\description{
Calculates the normalized (w.r.t. mean) root mean squares error
}
\usage{
RMSE(meas, pred)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{meas}{Measured values}
  \item{pred}{Predicted values}
}

